requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiogram_calendar==0.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp_socks==0.10.1
aiosignal==1.4.0
annotated-types==0.7.0
APScheduler==3.11.0
attrs==25.3.0
bcrypt==5.0.0
certifi==2025.8.3
cffi==2.0.0
cryptography==46.0.1
dotenv==0.9.9
frozenlist==1.7.0
idna==3.10
invoke==2.2.0
magic-filter==1.0.12
multidict==6.6.4
paramiko==3.4.0
propcache==0.3.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
psycopg2-binary==2.9.11
PyNaCl==1.6.0
python-dotenv==1.1.1
python-socks==2.7.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzlocal==5.3.1
yarl==1.20.1
